+++ processor/src/test/resources/expected/com/example/parameterized_type/Arez_ResolvedModel.java
@@ -1,0 +1,122 @@
+package com.example.parameterized_type;
+
+import java.util.Objects;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import org.realityforge.arez.Arez;
+import org.realityforge.arez.ArezContext;
+import org.realityforge.arez.Component;
+import org.realityforge.arez.Disposable;
+import org.realityforge.arez.Observable;
+import org.realityforge.braincheck.Guards;
+
+@Generated("org.realityforge.arez.processor.ArezProcessor")
+final class Arez_ResolvedModel extends ResolvedModel implements Disposable {
+  private static volatile long $$arez$$_nextId;
+
+  private final long $$arez$$_id;
+
+  private boolean $$arez$$_disposed;
+
+  @Nullable
+  private final ArezContext $$arez$$_context;
+
+  private final Component $$arez$$_component;
+
+  private final Observable<Boolean> $$arez$$_disposedObservable;
+
+  @Nonnull
+  private final Observable<Integer> $$arez$$_value;
+
+  Arez_ResolvedModel() {
+    super();
+    this.$$arez$$_context = Arez.areZonesEnabled() ? Arez.context() : null;
+    this.$$arez$$_id = $$arez$$_nextId++;
+    this.$$arez$$_component = Arez.areNativeComponentsEnabled() ? $$arez$$_context().createComponent( "ResolvedModel", $$arez$$_id(), $$arez$$_name(), null, null ) : null;
+    this.$$arez$$_disposedObservable = $$arez$$_context().createObservable( Arez.areNativeComponentsEnabled() ? this.$$arez$$_component : null, Arez.areNamesEnabled() ? $$arez$$_name() + ".isDisposed" : null, Arez.arePropertyIntrospectorsEnabled() ? () -> this.$$arez$$_disposed : null, null );
+    this.$$arez$$_value = $$arez$$_context().createObservable( Arez.areNativeComponentsEnabled() ? this.$$arez$$_component : null, Arez.areNamesEnabled() ? $$arez$$_name() + ".value" : null, Arez.arePropertyIntrospectorsEnabled() ? () -> super.getValue() : null, Arez.arePropertyIntrospectorsEnabled() ? v -> super.setValue( v ) : null );
+    if ( Arez.areNativeComponentsEnabled() ) {
+      this.$$arez$$_component.complete();
+    }
+  }
+
+  final ArezContext $$arez$$_context() {
+    return Arez.areZonesEnabled() ? this.$$arez$$_context : Arez.context();
+  }
+
+  final long $$arez$$_id() {
+    return this.$$arez$$_id;
+  }
+
+  String $$arez$$_name() {
+    return "ResolvedModel." + $$arez$$_id();
+  }
+
+  @Override
+  public boolean isDisposed() {
+    if ( $$arez$$_context().isTransactionActive() && !this.$$arez$$_disposedObservable.isDisposed() )  {
+      this.$$arez$$_disposedObservable.reportObserved();
+      return this.$$arez$$_disposed;
+    } else {
+      return this.$$arez$$_disposed;
+    }
+  }
+
+  @Override
+  public void dispose() {
+    if ( !isDisposed() ) {
+      this.$$arez$$_disposed = true;
+      if ( Arez.areNativeComponentsEnabled() ) {
+        this.$$arez$$_component.dispose();
+      } else {
+        $$arez$$_context().safeAction( Arez.areNamesEnabled() ? $$arez$$_name() + ".dispose" : null, () -> { {
+          this.$$arez$$_disposedObservable.dispose();
+          this.$$arez$$_value.dispose();
+        } } );
+      }
+    }
+  }
+
+  @Override
+  public Integer getValue() {
+    Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
+    this.$$arez$$_value.reportObserved();
+    return super.getValue();
+  }
+
+  @Override
+  public void setValue(final Integer value) {
+    Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
+    if ( !Objects.equals(value, super.getValue()) ) {
+      super.setValue(value);
+      this.$$arez$$_value.reportChanged();
+    }
+  }
+
+  @Override
+  public final int hashCode() {
+    return Long.hashCode( $$arez$$_id() );
+  }
+
+  @Override
+  public final boolean equals(final Object o) {
+    if ( this == o ) {
+      return true;
+    } else if ( null == o || !(o instanceof Arez_ResolvedModel) ) {
+      return false;
+    } else {
+      final Arez_ResolvedModel that = (Arez_ResolvedModel) o;;
+      return $$arez$$_id() == that.$$arez$$_id();
+    }
+  }
+
+  @Override
+  public final String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + $$arez$$_name() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}