+++ processor/src/test/resources/input/com/example/parameterized_type/ResolvedModel.java
@@ -1,0 +1,9 @@
+package com.example.parameterized_type;
+
+import org.realityforge.arez.annotations.ArezComponent;
+
+@ArezComponent
+class ResolvedModel
+  extends ParentModel<Integer>
+{
+}