+++ processor/src/test/resources/input/com/example/parameterized_type/UnresolvedModel.java
@@ -1,0 +1,9 @@
+package com.example.parameterized_type;
+
+import org.realityforge.arez.annotations.ArezComponent;
+
+@ArezComponent
+class UnresolvedModel<X extends Number>
+  extends ParentModel<X>
+{
+}